REST based serverShellCommands
