+++ package.json
@@ -148,7 +148,6 @@
     "bcrypt": "^5.0.1",
     "bowser": "^1.9.4",
     "cheerio": "1.0.0-rc.9",
-    "chokidar": "^3.5.0",
     "classnames": "^2.2.6",
     "cloudinary": "^1.23.0",
     "cloudinary-core": "2.5.0",
