+++ packages/lesswrong/server/serverStartup.ts
@@ -12,10 +12,8 @@ import { getCommandLineArguments } from './commandLine';
 import { startWebserver } from './apolloServer';
 import { initGraphQL } from './vulcan-lib/apollo-server/initGraphQL';
 import { createVoteableUnionType } from './votingGraphQL';
-import { Globals, Vulcan } from '../lib/vulcan-lib/config';
+import { Globals } from '../lib/vulcan-lib/config';
 import process from 'process';
-import chokidar from 'chokidar';
-import fs from 'fs';
 
 async function serverStartup() {
   // eslint-disable-next-line no-console
@@ -97,7 +95,6 @@ async function serverStartup() {
     initShell();
   } else {
     if (!isAnyTest) {
-      watchForShellCommands();
       // eslint-disable-next-line no-console
       console.log("Starting webserver");
       startWebserver();
@@ -161,49 +158,6 @@ function initShell()
   });
   r.context.Globals = Globals;
   r.context.Vulcan = Globals;
-}
-
-const compileWithGlobals = (code: string) => {
-  // This is basically just eval() but done in a way that:
-  //   1) Allows us to define our own global scope
-  //   2) Doesn't upset esbuild
-  const callable = (async function () {}).constructor(`with(this) { await ${code} }`);
-  const scope = {Globals, Vulcan};
-  return () => {
-    return callable.call(new Proxy({}, {
-      has () { return true; },
-      get (target, key) {
-        if (typeof key !== "symbol") {
-          return global[key] ?? scope[key];
-        }
-      }
-    }));
-  }
-}
-
-// Monitor ./tmp/pendingShellCommands for shell commands. If a JS file is
-// written there, run it then delete it. Security-wise this is okay because
-// write-access inside the repo directory is already equivalent to script
-// execution.
-const watchForShellCommands = () => {
-  const watcher = chokidar.watch('./tmp/pendingShellCommands');
-  watcher.on('add', async (path) => {
-    const fileContents = fs.readFileSync(path, 'utf8');
-    // eslint-disable-next-line no-console
-    console.log(`Running shell command: ${fileContents}`);
-    fs.unlinkSync(path);
-    try {
-      const func = compileWithGlobals(fileContents);
-      const result = await func();
-      // eslint-disable-next-line no-console
-      console.log("Finished. Result: ", result);
-    } catch(e) {
-      // eslint-disable-next-line no-console
-      console.log("Failed.");
-      // eslint-disable-next-line no-console
-      console.log(e);
-    }
-  });
 }
 
 void serverStartup();
