+++ scripts/serverShellCommand.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
-# Run Javascript in an existing server process. This works by putting the script
-# into ./tmp/pendingShellCommands, where the server will detect it, run it then
-# delete it. (See watchForShellCommands in serverStartup.ts)
+# Run Javascript in an existing server process - note that this is disabled on
+# prod and staging servers for security reasons
 
-WAIT=0
+WAIT=false
 if [ "$1" = "--wait" ]; then
-	WAIT=1
+	WAIT=true
 	shift
 fi
 
@@ -13,13 +12,6 @@ COMMAND="$1"
 echo "Checking that a local server is running..."
 scripts/waitForServer.sh
 
-OUTFILE=tmp/pendingShellCommands/command$$.js
-
-mkdir -p tmp/pendingShellCommands
-echo "$COMMAND" >$OUTFILE
-
-if [ $WAIT = 1 ]; then
-	while test -f "$OUTFILE"; do
-		sleep 0.2
-	done
-fi
+curl -X POST http://localhost:3000/api/serverShellCommand \
+	-d "{\"command\": \"$COMMAND\", \"wait\": $WAIT}" \
+	-H "Content-Type: application/json"
